Changed event struct for LogLock and LogBurn to use uint8 for decimals instead of uint256.

+++ smart-contracts/contracts/BridgeBank/EthereumBankStorage.sol
@@ -2,40 +2,39 @@
 pragma solidity 0.8.0;
 
 contract EthereumBankStorage {
-
     /**
-    * @notice current lock and or burn nonce
-    */
+     * @notice current lock and or burn nonce
+     */
     uint256 public lockBurnNonce;
 
     /*
-    * @notice {DEPRECATED}
-    */
+     * @notice {DEPRECATED}
+     */
     mapping(address => uint256) private lockedFunds;
 
     /*
-    * @notice map the token symbol to the token address
-    */
+     * @notice map the token symbol to the token address
+     */
     mapping(string => address) public lockedTokenList;
 
     /**
-    * @notice contract decimals based off of contract address
-    */
-    mapping (address => uint8) public contractDecimals;
+     * @notice contract decimals based off of contract address
+     */
+    mapping(address => uint8) public contractDecimals;
 
     /**
-    * @notice contract symbol based off of address
-    */
-    mapping (address => string) public contractSymbol;
+     * @notice contract symbol based off of address
+     */
+    mapping(address => string) public contractSymbol;
 
     /**
-    * @notice contract name based off of address
-    */
-    mapping (address => string) public contractName;
+     * @notice contract name based off of address
+     */
+    mapping(address => string) public contractName;
 
     /*
-    * @notice gap of storage for future upgrades
-    */
+     * @notice gap of storage for future upgrades
+     */
     uint256[97] private ____gap;
 
     /*
@@ -47,7 +46,7 @@ contract EthereumBankStorage {
         address _token,
         uint256 _value,
         uint256 indexed _nonce,
-        uint256 _decimals,
+        uint8 _decimals,
         uint256 _networkDescriptor
     );
 
@@ -57,15 +56,11 @@ contract EthereumBankStorage {
         address _token,
         uint256 _value,
         uint256 indexed _nonce,
-        uint256 _decimals,
+        uint8 _decimals,
         string _symbol,
         string _name,
         uint256 _networkDescriptor
     );
 
-    event LogUnlock(
-        address _to,
-        address _token,
-        uint256 _value
-    );
+    event LogUnlock(address _to, address _token, uint256 _value);
 }
